+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+
+## 0.2.0
+August 21, 2022
+- import '../example/test/widget_test.dart';
+- Supply continuous integration testing
+
 ## 0.1.0
 July 18, 2022
 - Initial Development Release